Add files via upload
The dining out behaviour of consumers in India has seen a marked transformation in recent years. The food services industry is one of the most promising sectors of investment. In this project we explore the interests of the Indian consumer in cuisines, popular dishes and their expectations from restaurants. We do this by understanding the factors affecting the prices of food in restaurants. We use decision tree and random forests to understand the various factors influencing the prices such as city, rating, cuisine and services provided by the restaurant. Our findings show that in general, people prefer dishes to be in a reasonable price range of 100 to 400 rupees. The famous tourist cities and the cities that are ranked among higher tiers tend to have higher food prices. We explore the popular cuisines served in restaurants with the consumer preferences and understand their impact on restaurant food prices.